+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.2.8] - XXXX-XX-XX
+## [0.2.8] - 2021-03-05
 
 ### Added
 - `ContainsEdgePass` added before `ReachingDefPass`